refactor/reflow: split tests into separate file
